src/app/shared/components/movie-card/movie-card.component.ts
import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
import { Movie } from '../../models/movies-response.model';

@Component({
  selector: 'app-movie-card',
  templateUrl: './movie-card.component.html',
  styleUrls: ['./movie-card.component.scss'],
  changeDetection: ChangeDetectionStrategy.OnPush
})
export class MovieCardComponent {
  @Input() movie: Movie = {} as Movie;
  @Output() viewDetail = new EventEmitter<Movie>()
  image_url: string = "https://image.tmdb.org/t/p/original";
}
